pyproject.toml
[tool.poetry]
name = "Message Queue API"
version = "0.1.0"
description = ""
authors = ["Payon <ken71301@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.0"
uvicorn = "0.31.0"
requests = "^2.32.0"
asgi-lifespan = "^1.0.1"
httpx = "^0.27.0"
rq = "^1.16.0"
rq-dashboard-fast = "^0.5.9"
pydantic-settings = "^2.5.2"
fakeredis = "^2.25.0"
ruff = "^0.6.9"


[tool.poetry.dev-dependencies]
pytest = "^8.2"
pytest-mock = "^3.14.0"
pre-commit = "^2.6.0"
tox = "^3.19.0"
pytest-cov = "^5.0.0"
pytest-asyncio = "^0.23.7"
ipdb = "^0.13.9"
yapf = "^0.31.0"

[tool.pytest.ini_options]
pythonpath = [
  "."
]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
